update collision functions
update intersect shape(almost fixed, 2 more tests to fix)

fix direct space functions

fix max contacts count

upd

fix type conversion to match current godot space

fix density and space transform

refactor

fix pointer crash

add manifold code

update collision normal

fix intersect_shape tests

add shape collide impl

add impl for safe/unsafe fractions

apply b2Common hack

fix macos issue

update path of lib to addons/physics_server_box2d/

fix typo

rename artifactts

use wildcard pattern for addons folder

fix normal being zero

+++ src/servers/physics_server_box2d.h
@@ -43,7 +43,7 @@ class PhysicsServerBox2D : public PhysicsServer2DExtension {
 	mutable RID_PtrOwner<Box2DArea, true> area_owner;
 	mutable RID_PtrOwner<Box2DBody, true> body_owner;
 	mutable RID_PtrOwner<Box2DJoint, true> joint_owner;
-	float step_amount = 0;
+	double step_amount = 0;
 
 	RID _shape_create(ShapeType p_shape);
